requirements.txt
requests~=2.31.0
pandas~=2.2.1
numpy~=1.26.4
tqdm~=4.66.2
openpyxl~=3.1.2
open_rarity~=0.7.5
seaborn~=0.13.2
matplotlib~=3.10.0
ipython~=8.31.0
python-binance~=1.0.27
scipy~=1.15.1